requirements.txt
datasets==2.2.2
pytorch-lightning==1.6.4
PyYAML==6.0
scikit-learn==1.1.2
transformers==4.18.0
wandb==0.12.18